refactor/#46 util 함수 생성 및 적용
