+++ src/utils/format.date.js
@@ -1,0 +1,11 @@
+export default function formatDate(createAt) {
+  const date = new Date(createAt);
+
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  const hour = date.getHours();
+  const min = date.getMinutes();
+
+  return `${year}년 ${month}월 ${day}일 ${hour}시 ${min}분`;
+}
